Fix sendfile for files >2GB on 64bit architectures Fix >2GB sendfile on XFS : ----------------------------------------------------------------------
suse-commit: 73d1b9be98529c8ec6528a9841283ab1fc793c75

+++ fs/xfs/linux/xfs_file.c
@@ -261,7 +261,7 @@ linvfs_sendfile(
 	void			*target)
 {
 	vnode_t			*vp = LINVFS_GET_VP(filp->f_dentry->d_inode);
-	int			error;
+	long			error;
 
 	VOP_SENDFILE(vp, filp, ppos, 0, count, actor, target, NULL, error);
 	return error;
